支持Segment之间资金的内部互转和换汇下单

+++ TigerOpenAPI/Quote/Model/FutureTickModel.cs
@@ -17,7 +17,7 @@ namespace TigerOpenAPI.Quote.Model
     public Int64 EndIndex { get; set; }
 
     [JsonProperty(PropertyName = "limit")]
-    public Int16 Limit { get; set; } = 200;
+    public Int32 Limit { get; set; } = 200;
 
     public FutureTickModel() : base()
     {
